Countdown Timer for PlanetScene
a timer has been implemented that counts down an amount of seconds (default: 315) until executing a function "excitingThing()" which will be replaced with another function to handle the end of the exploration/extraction gameplay phase of the game. The timer also displays the amount of seconds left - in the form of Minutes:Seconds - on a default text UI that can (and should) be moved or changed later.